auth: split login and verify tasks
